requirements.txt
telethon
python-decouple